chore(sync): differentiate the insert_id and user_id hmac keys
The HMAC key we use to hash user_id is super-secret and only available
inside the prod IAM. Since we don't have access to that in dev, where
the Sync import runs, use a different environment variable name to make
it more obvious that the two keys are different. This doesn't make any
functional difference to the data we're sending to Amplitude, but it may
prevent confusion in the future (especially after the marketing import
script lands).