Add support for routing 'identify' events to the batch API.
This change allows the 'identify' amplitude events to be split between
the identify API and the batch API.

The BATCH_API_PERCENTAGE env var (int 0 - 100, default 75) is used to
decide which API to use to handle a particular event.

Other new config values added in this patch:

* BATCH_API_MAX_EVENTS_PER_BATCH, number of events per batch API
submission (default 10)

* BATCH_API_WORKER_COUNT, number of batch API queue workers per
fxa-amplitude-send instance (default 1)

Fixes #107.